improve storybook examples
